+++ packages/agent-core-v2/src/index.ts
@@ -438,10 +438,6 @@ export * from '#/agent/task/taskOps';
 export * from '#/agent/task/taskService';
 import '#/features/cron/configSection';
 export * from '#/features/cron/cronTask';
-export * from '#/features/cron/cron-expr';
-export * from '#/features/cron/format';
-export * from '#/features/cron/jitter';
-export * from '#/features/cron/clock';
 export * from '#/features/cron/configSection';
 export * from '#/features/cron/cronAgentRuntime';
 export * from '#/features/cron/cronOps';
