+++ packages/agent-core-v2/test/features/cron/cron-expr.test.ts
@@ -5,7 +5,7 @@ import {
   cronToHuman,
   hasFireWithinYears,
   parseCronExpression,
-} from '#/features/cron/cron-expr';
+} from '#/features/cron/internal/cron-expr';
 
 function localDate(
   year: number,
