+++ packages/agent-core-v2/test/features/cron/jitter.test.ts
@@ -1,11 +1,11 @@
 import { describe, expect, it } from 'vitest';
 
-import { parseCronExpression } from '#/features/cron/cron-expr';
+import { parseCronExpression } from '#/features/cron/internal/cron-expr';
 import {
   DEFAULT_CRON_JITTER_CONFIG,
   jitteredNextCronRunMs,
   oneShotJitteredNextCronRunMs,
-} from '#/features/cron/jitter';
+} from '#/features/cron/internal/jitter';
 
 function localDate(
   year: number,
